kvserver: move SideloadStorage to a package
This commit moves and exports the SideloadStorage interface and its
implementation to the new logstore package. The package will accumulate types
forming the separated log storage.

Only the unit tests which did not depend on Replica are moved to the new
package. Tests which depended on Replica stay in kvserver for now, with a TODO
to untangle them from Replica and move away. They might stay too because they
test integration between kvserver and logstore.

Release note: None
